Move `String+strippedHTML` to Networking layer for reuse.
